third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-1/edit-media-text.html

<html>
<head>
<style>
@media screen and (max-device-width: 100000px) {
    #inspected {
        color: green;
    }
    #inspected {
        color: blue;
    }
}
@media screen and (max-device-width: 200000px) {
    #other {
        color: green;
    }
}
</style>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function test() {
  ElementsTestRunner.selectNodeAndWaitForStyles('inspected', step1);

  function step1() {
    TestRunner.addResult('=== Before media text modification ===');
    ElementsTestRunner.dumpSelectedElementStyles(true);
    var section = ElementsTestRunner.firstMatchedStyleSection();
    var mediaTextElement = ElementsTestRunner.firstMediaTextElementInSection(section);
    mediaTextElement.click();
    mediaTextElement.textContent = 'screen and (max-device-width: 99999px)';
    ElementsTestRunner.waitForMediaTextCommitted(step2);
    mediaTextElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));
  }

  function step2() {
    TestRunner.addResult('=== After valid media text modification ===');
    ElementsTestRunner.dumpSelectedElementStyles(true);
    var section = ElementsTestRunner.firstMatchedStyleSection();
    var mediaTextElement = ElementsTestRunner.firstMediaTextElementInSection(section);
    mediaTextElement.click();
    mediaTextElement.textContent = 'something is wrong here';
    ElementsTestRunner.waitForMediaTextCommitted(step3);
    mediaTextElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));
  }

  function step3() {
    TestRunner.addResult('=== After invalid media text modification ===');
    ElementsTestRunner.dumpSelectedElementStyles(true);
    TestRunner.completeTest();
  }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that editing media text updates element styles.
</p>

<div id="inspected" style="color: red">Text</div>
<div id="other"></div>

</body>
</html>
