third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-1/edit-media-text-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that editing media text updates element styles.

=== Before media text modification ===
[expanded] 
element.style { ()
    color: red;

[expanded] 
@media screen and (max-device-width: 100000px)
#inspected { (<style>…</style>)
/-- overloaded --/     color: blue;

[expanded] 
@media screen and (max-device-width: 100000px)
#inspected { (<style>…</style>)
/-- overloaded --/     color: green;

[expanded] 
div { (user agent stylesheet)
    display: block;

=== After valid media text modification ===
[expanded] 
element.style { ()
    color: red;

[expanded] 
@media screen and (max-device-width: 99999px)
#inspected { (<style>…</style>)
/-- overloaded --/     color: blue;

[expanded] 
@media screen and (max-device-width: 99999px)
#inspected { (<style>…</style>)
/-- overloaded --/     color: green;

[expanded] 
div { (user agent stylesheet)
    display: block;

=== After invalid media text modification ===
[expanded] 
element.style { ()
    color: red;

[expanded] 
#inspected { (<style>…</style>)
/-- overloaded --/     color: blue;

[expanded] 
#inspected { (<style>…</style>)
/-- overloaded --/     color: green;

[expanded] 
div { (user agent stylesheet)
    display: block;
